// packages/sdk/src/enums/ShoyuError.ts
export enum ShoyuError {
  ORDER_EXPIRED = "ORDER_EXPIRED",
  INVALID_MERKLE_ROOT = "INVALID_MERKLE_ROOT",
  INVALID_CHAINID = "INVALID_CHAINID",
  INVALID_VERIFYING_CONTRACT = "INVALID_VERIFYING_CONTRACT",
  INVALID_NFT_TOKENID = "INVALID_NFT_TOKENID",
  INVALID_NFT_TOKEN_AMOUNT = "INVALID_NFT_TOKEN_AMOUNT",
  INVALID_PROTOCOL_FEE = "INVALID_PROTOCOL_FEE",
  INVALID_TRADE_DIRECTION = "INVALID_TRADE_DIRECTION",
  NATIVE_TOKEN_ONLY = "NATIVE_TOKEN_ONLY",
  WRAPPED_NATIVE_TOKEN_ONLY = "WRAPPED_NATIVE_TOKEN_ONLY",
}